fix solr cookbooks
the cookbooks fetched some files from a git repository
through a web-interface. Needless to say that this was
a bad idea in the first place.

Since it broke now for a second time, I just add the
resources to that cookbook, because I asume they do not
change anyway (might of course be, that I'm wrong)

Change-Id: I6385e150a15b0f165fe75b08e69708eb4fe763fe